refactor(tools): consolidate subprocess pipeline, file I/O, and context into shared modules
- Extract _SHELL_CONTROL, _drain, _decode_output, _MAX_OUTPUT_BYTES, and
  _subprocess_sync into lionagi/tools/_subprocess.py; tools/code/bash.py,
  tools/coding.py, and agent/permissions.py all import from the canonical source

- coding.py _read_file_sync and _list_dir_sync now delegate to file/reader.py
  (_read_sync, _list_dir_sync) gaining symlink safety and _MAX_LIST_FILES cap;
  _edit_file_sync uses _write_text_no_follow from file/editor.py for symlink-safe
  writes; _write_file_sync unchanged but imports consolidated

- coding.py context closure replaced by ContextTool().bind(branch) delegation;
  removes duplicate _ensure_current_progression and ~110 lines of context logic;
  agents now get restore/compact actions for free; files_tracked injected on status

- Update tests to import from canonical locations and patch at _subprocess module
